requirements.txt
requests==2.31.0			# imported as "requests"
pathlib==1.0.1 				# imported as "pathlib"
urllib3==2.0.7				# imported as "urllib"
python-dotenv==1.0.0	# imports as "dotenv"